requirements.txt
streamlit>=1.28.0
langchain-core
langchain
langchain-openai
langgraph
python-dotenv